Implement ability to obtain stack from yjit exits
NOTE: this is still WIP and needs to be rebased with master and
reimplemented in YJIT Rust. I've opened this PR for demo purposes.

Using Stackprof we can get more information about where a particular
exit is coming from. When tenderlove and I were working on finding some
YJIT exits we noticed that Rails has a lot of exits for
`getinstancevariable`. While we could get the exit reason it was unclear
what code specifically was exiting. From there we implemented the
ability to obtain the stack trace and feed it into Stackprof so that
we're able to find the exact code for an exit. This information is very
useful for prioritizing improvements to areas YJIT is exiting frequently
by providing files and line numbers of the exited code.

To do:

* needs docs
* needs tests?

Co-authored-by: Aaron Patterson <tenderlove@ruby-lang.org>