Fix Range#count when beg and end are Numeric
In https://bugs.ruby-lang.org/issues/18984 a bug is described when calling
`Range#size` on a range where the beginning and end are Numberic, a value
is returned. However when calling `Range#count` a `TypeError` is raised.
A comment on the report felt that the bug was in `Range#size` and that should
raise a `TypeError` as well. However, looking at the code I don't think this
is correct. `Range#count` can be an enumerable or it can behave the same as
`Range#size`. The way I fixed this bug was by checking if we have a
`Numeric` value for the beginning and end of the range, AND we don't
have a block, we should treat `Range#count` the same as `Range#size` and
not raise a `TypeError`. While the original report was using
`(0.51..5.quo(2))` for the example Range, I found this bug was also
present wihtout using `quo` so the tests reflect that.

I duplicated all the tests from `test_size` in order to ensure that
expected behavior is tested for `Range#count` and it's easier to see the
difference between their behavior.

Fixes: [ruby-core:18984]