Initialize ivars to `nil`
This branch initializes all ivar's to `nil`. By doing this we create
less new shapes becasue the ivar order is already set. This branch is
experimental. One caveat is this will break code that checks if an
instance variable is defined because it will always be
"instance-variable" after this change.