Refactor update location functions
In about half the cases that `rb_gc_impl_location` was called, there was
always a check for `rb_gc_impl_object_moved_p`. In this refactor I made
a new function `rb_gc_impl_moved_object_location` that does the
`rb_gc_impl_object_moved_p` check and the callers. I also made this
change in the `TYPED_UPDATE_IF_MOVED` macro. The change required
`rb_gc_impl_object_moved_p` to be moved up. The goal of this refactor is
to make the code more concise, while also providing a public API for
other GC implementations to not need to perform this conditional.